requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
duckduckgo_search==8.0.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.0
groq==0.22.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
kafka==1.3.5
kafka-python==2.1.5
langchain==0.3.23
langchain-core==0.3.52
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langgraph==0.3.30
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.31
lxml==5.3.2
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.4
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
phi==0.6.7
phidata==2.7.10
primp==0.14.0
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.1.2
tomli==2.2.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
zstandard==0.23.0
Flask==3.0.2
